threads-auto/threads-auto/requirements.txt
fastapi
uvicorn
line-bot-sdk
playwright
psycopg2-binary
python-dotenv